requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.0

# LLM Integration
groq==0.4.2
httpx==0.25.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment Management
python-dotenv==1.0.0

# Document Processing (for RAG)
pypdf2==3.0.1
pypdf==4.0.1
python-docx==1.1.0

# Utilities
tiktoken==0.5.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0